refactor parsing into own class, and made constructor for State
