+++ src/main/java/Duke.java
@@ -1,3 +1,4 @@
+import com.util.json.JsonParser;
 import com.util.json.SaveData;
 import java.util.Scanner;
 import java.util.NoSuchElementException;
@@ -13,8 +14,7 @@ public class Duke {
         // greet
 
         Scanner scanner = new Scanner(System.in);
-        State state = new State();
-        state.list = SaveData.parseJsonFile(SaveData.read());
+        State state = new State(JsonParser.parseJsonFile(SaveData.read()));
         try {
             while (!state.toExit) {
                 String input = scanner.nextLine();
